chore(badges): Enhance warning in readme and describe how to allow configuration reading from frontend. Move remaining tests files beside the tested code
